pipelines/rollback.py
import boto3
import yaml

# Placeholder: load config
with open('configs/config.yaml') as f:
    config = yaml.safe_load(f)

# Placeholder: get current and previous model accuracy
current_acc = 0.65  # Simulate drop
previous_acc = 0.75
threshold = config['model']['ab_test_threshold']

if previous_acc - current_acc > threshold:
    print("Accuracy drop detected! Rolling back to previous model...")
    # Placeholder: add logic to update ECS service or S3 pointer
else:
    print("Model performance is acceptable. No rollback needed.") 